libknet: add seq_num to sender host
bcast for packets that needs to be sent to all nodes
ucast for direct packets

the difference is necessary to avoid unncessary re-encryption
of packets. This will make more sense once packet inspector
is in place, but basically seq_num are per host and not all
packets will go everywhere, creating undesirable holes in the
packet_ring_buffer, that in turns, in a non-encrypted network
could lead to some annoyance.

Signed-off-by: Fabio M. Di Nitto <fdinitto@redhat.com>