[send/recv] Fix handling of HOSTINFO packets
HOSTINFO packets are nothing more than special cased DATA packets.

The contents of HOSTINFO is stored inside DATA of DATA packets and
up to the point where we need to use the HOSTINFO data, we need
to treat the packets as DATA for defragmentation and deduplication.

This also fixes a problem when using round-robin and active switching
policy where the HOSTINFO exchange would loop forever in a storm
because duplicate pckts were being handled at once.

Signed-off-by: Fabio M. Di Nitto <fdinitto@redhat.com>