[PMTUd] Drop concept of link_mtu and fix data mtu exported values (pass 1)
Cleanup a poor attempt to export all MTU data to the users.

The original idea was to provide information about real link onwire mtu
and max user data mtu (that could be sent on wire without fragmentation)
that adjusted the value for IP protocol, knet header and crypto (if configured).

Effectively there is no use for the real link onwire mtu and it only involves
extra complexity to export lots of extra data to the end user that would have
to redo the calculations internally to gather data mtu.

Also fix the status.mtu value to match knet view of every link.

Signed-off-by: Fabio M. Di Nitto <fdinitto@redhat.com>