[handle] Allow threads to recognized shutdown before stopping them.
On busy threads and shutdown, there is a chance that the thread
didn't have time to process the shutdown requests and even tho
we are closing down, some locks can be destroyed while still locked.

In theory we could have each thread report that it has ack'ed the shutdown
and then close, but a sleep seems to be more than good enough for now.

Signed-off-by: Fabio M. Di Nitto <fdinitto@redhat.com>