Add libnss iovec PoC
this is a small modified version of the code example available here:

https://developer.mozilla.org/en-US/docs/Mozilla/Projects/NSS/nss_sample_code/NSS_Sample_Code_sample2

to show that it is possible to encrypt from multiple src buffers
into a final buffer.

this approach will allow us to remove a lot of memcpy around the
recv/send thread and replace them with iovec mapping all across.

Possible side effects:
- less memory consumption (mapping vs allocating)
- more complex code / crypto (internal) API changes