[link] switch link_get_status to write lock
previous to the introduction of detailed link stats, the values
contained in the status structure would hardly change over time
making a rdlock more than enough to get the right data.

Now that the values are updated constantly, we need to remove
the window of thread X writing values and thread Y to read them
while being written.

Signed-off-by: Fabio M. Di Nitto <fdinitto@redhat.com>