libknet: Optimised path for local traffic
In corosync I bypassed knet for local traffic by sending to datafd+1,
this adds that functionality back into knet to avoid exceptions in
client code and allow us to keep statistics on local traffic.

This only happens if link 0 on the local host is enabled and can be
disabled by disabling that 'link'. There is no actual link of course
just a short-cut send.

Signed-off-by: Christine Caulfield <ccaulfie@redhat.com>