[build] release.mk: reflect current release publishing practice
Core author(s) had previous affiliation with project hosting on
fedorahosted.org,  but since the sunset of that hosting[1], it's no
longer an option.

Unfortunately, publish-by-scp convenience of fedorahosted.org won't fly
with GitHub that insists on the (primarily REST) API to get things
automated, so update "publish" recipe to that effect, providing some
hints on possible adaption of the process.

Also, turn "git push --tags" to a bit safer "git push --follow-tags"
for good measure.

[1] https://lists.fedoraproject.org/archives/list/announce@lists.fedoraproject.org/message/BWSMCGZPPNG3JOCFQ6Z74MIBU7FG3KGB/

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>