[PMTUd] drop (now) unnecessary and dangerous usleep
prior to all threads being able to notify PMTUd of EMSGSIZE errors,
we had this random usleep in there to have time to collect data.
It was working at the time, but it's a bad idea.

On super large clusters (>66 nodes) with 4 links on each node, when
applying heavy load (cpghum on all nodes at once), the average latency
between nodes can increase so much that the PMTUd thread usleep
could literally block corosync for seconds at a time.

Drop the usleep and live happily ever after

Signed-off-by: Fabio M. Di Nitto <fdinitto@redhat.com>