Use LD_LIBRARY_PATH to find the build tree artifacts and disable RPATH across the project
The libtool wrapper scripts of the test binaries set LD_LIBRARY_PATH to
find the libknet library and its modules in the build tree, but RPATH
overrides this setting.  This is why RPATH is deprecated, so we switch
to RUNPATH instead by using --enable-new-dtags, which is already the
linker default on Debian for example.

Based on the original patch from Ferenc.

Signed-off-by: Ferenc Wágner <wferi@debian.org>
Signed-off-by: Fabio M. Di Nitto <fdinitto@redhat.com>