build: sanitize LDFLAGS handling
The Automake manual states that according to the GNU Coding Standards
the so-called "Variables reserved for the user" mustn't be changed by
the build system, and they must override the default settings from the
build system.  The "Flag Variables Ordering" section provides the
recipes on which we build here.

Using $lt_prog_compiler_pic directly shouldn't be necessary, let's try
leaving it out.

Signed-off-by: Ferenc Wágner <wferi@debian.org>