[tx/rx] improve error checking when running external code
libknet cannot guarantee that the dst_host_filter is sane nor that it cannot
be exploited to return garbage to knet.

there is an infinitesimal possibility that, if the plugin returns total crap,
knet could crash by trying to access out-of-bound memory.

Prevent that by adding specific checks both in TX/RX.

Signed-off-by: Fabio M. Di Nitto <fdinitto@redhat.com>