[build] improve compiler parameter checking
in the old way of checking compiler extra warnings, we would
check one warning at a time, but it doesn't really detect
incompatiblities between parameters till we attempt to build
the first object.

pass EXTRA_WARNINGS to the function that does check for parameters
in a cumulative fashion, so that potential incompatibilities can
be detected at configure time vs build time.

Signed-off-by: Fabio M. Di Nitto <fdinitto@redhat.com>