[transports] fix usage after free of transport info memory
spotted while implementing functional testing for rekey.

when configuring multiple hosts (3+) with 1 link sharing
the same listener, but NOT enabling the link, if the user
attempts to shutdown knet_h, the first call to knet_link_clear_config
will not recognize that the listener is still in use by another
link and will release all resources.
At the time of invoking clear_config on the second link, transport_clear_config
will access already freed memory, or alternatively, a call to
link_enable would cause a crash because the listener does no
longer exist.

Signed-off-by: Fabio M. Di Nitto <fdinitto@redhat.com>