doxyxml: Tidy descriptions of functions as parameters
If a complex function pointer was passed as a parameter then doxyxml
tryied to line up all the other parameters with it - making a mess
by having lots of blank space between the type and the name.

Now we enforce a maximum type length (a line-ish) so that shorter
tyopes will line up OK and the really long ones will be left to their
own devices.