man: Tidy manpages (#215)
* man: Tidy manpages for libnozzle

doxygen works in mysterious ways, adding a blank line before
@brief makes the lines following that much tidier.

So now instead of

nozzle_close nozzle - pointer to the nozzle struct to destroy

we get:

nozzle_close

       nozzle - pointer to the nozzle struct to destroy

* doxyxml: Cope with pointers-to-pointers passed as params

Double pointers showed as ' * *name' when they should be '  **name'.

Also tidy STRUCTURES display so that they are not indented too much,

* man: Similar @brief fixes for libknet.h

* doxyxml: Tidy descriptions of functions as parameters

If a complex function pointer was passed as a parameter then doxyxml
tryied to line up all the other parameters with it - making a mess
by having lots of blank space between the type and the name.

Now we enforce a maximum type length (a line-ish) so that shorter
tyopes will line up OK and the really long ones will be left to their
own devices.