[sctp] check socket errors only after socket re-create process
For some unknown reasons (yet), this code used to work with older kernels,
but it recently broke.

Regardless of the kernel change, it is incorrect to get the socket status
before re-creating the socket and rely on the old status for the new
socket behavior.

By moving the check after the socket re-creation, then knet behaves correctly.

This patch will be merged once we have a full explanation of what changed
in kernel.

Original-Patch-by: Xin Long <lxin@redhat.com>
Signed-off-by: Fabio M. Di Nitto <fdinitto@redhat.com>