[lib] Don't use -rpath
rpath seems, understandably, to annoy some analysis tools. We
use it to locate the crypto & compression plugins but libknet also needs
to use the plugins from the build-tree when running tests.

A simple way to do this is have the plugin directory as part of the knet
handle then we can override it in the test suite, which already does
'illegal' accesses to some handle fields.

OK, it's not the MOST elegant way of doing this perhaps, but it is
simple and effective.