Sanitise sockaddr_storage structs coming in the API
Before we used to copy the whole sockaddr_storage from
the API into our structs. This meant that any junk left
at the end by the caller got into them and could
cause trouble with some comparisons.

This patch zeros out the extra bits of the sockaddr_storage
that are not valid.