[host] fix dst_seq_num initialization race condition
There is a potential race condition where the sender
is overloaded, sending data packets before pings
can kick in and set the correct dst_seq_num.

if this node is starting up (dst_seq_num = 0),
it can start rejecing valid packets and get stuck.

Set the dst_seq_num to the first seen packet and
use that as reference instead.

Signed-off-by: Fabio M. Di Nitto <fdinitto@redhat.com>