pmtud: Reset/restart pmtud when a node joins
If a new node joins with a 'black hole' that sets the effective
MTU to be less than the hardware MTU then it will fail to join.

So this patch restarts the pMTU processes each time a new node
joins so it has a chance to declare it's real MTU in time.